refactor: remove generateLRC function and its references from LyricEditor and LyricStudioPage components
