+++ src/components/lyric-studio/lyric-header.tsx
@@ -3,6 +3,7 @@ import { CardHeader, CardTitle } from '../ui/card';
 import type { LyricLine } from './lyric-line-item';
 import { Button } from '../ui/button';
 import { useAppContext } from '@/hooks/use-app-context';
+import { formatLRCTimestamp } from '@/lib/utils';
 
 interface LyricMetadata {
 	title: string;
@@ -21,7 +22,6 @@ export interface LRCData {
 export function LyricHeader({
 	showPreview,
 	setShowPreview,
-	generateLRC,
 	hasEmptyLyricLines,
 	lyricLines,
 	onShowExternalLyrics,
@@ -29,13 +29,59 @@ export function LyricHeader({
 }: {
 	showPreview: boolean;
 	setShowPreview: (show: boolean) => void;
-	generateLRC: () => LRCData;
 	hasEmptyLyricLines: () => boolean;
 	lyricLines: LyricLine[];
 	showExternalLyrics: boolean;
 	onShowExternalLyrics: () => void;
 }) {
 	const { areLyricLinesWithoutTimestamps } = useAppContext();
+
+	const generateLRC = () => {
+		// Sort lyrics by timestamp to ensure proper order
+		const sortedLyrics = [...lyricLines].sort(
+			(a, b) => (a.timestamp ?? 0) - (b.timestamp ?? 0)
+		);
+
+		const lrcData: LRCData = {
+			metadata: {
+				title: 'Untitled Song',
+				artist: 'Unknown Artist',
+				album: 'Unknown Album',
+				timestamps: sortedLyrics.map((line) => ({
+					time: formatLRCTimestamp(line?.timestamp ?? 0),
+					text: line.text,
+				})),
+			},
+		};
+
+		console.log('LRC Data (JSON format):', lrcData);
+		return lrcData;
+	};
+
+	const handleDownload = () => {
+		const lrcData = generateLRC();
+
+		// Generate LRC content
+		let lrcContent = `[ti:${lrcData.metadata.title}]\n`;
+		lrcContent += `[ar:${lrcData.metadata.artist}]\n`;
+		lrcContent += `[al:${lrcData.metadata.album}]\n\n`;
+
+		// Add timestamp lines
+		lrcData.metadata.timestamps.forEach(({ time, text }) => {
+			lrcContent += `[${time}]${text}\n`;
+		});
+
+		// Create and trigger download
+		const blob = new Blob([lrcContent], { type: 'text/plain' });
+		const url = URL.createObjectURL(blob);
+		const a = document.createElement('a');
+		a.href = url;
+		a.download = 'lyrics.lrc';
+		document.body.appendChild(a);
+		a.click();
+		document.body.removeChild(a);
+		URL.revokeObjectURL(url);
+	};
 	return (
 		<CardHeader className="flex flex-row items-center justify-between py-8 border-b">
 			<CardTitle className="flex items-center gap-2">
@@ -69,7 +115,7 @@ export function LyricHeader({
 					{showPreview ? 'Hide Preview' : 'Preview'}
 				</Button>
 				<Button
-					onClick={generateLRC}
+					onClick={handleDownload}
 					disabled={
 						hasEmptyLyricLines() ||
 						lyricLines.length === 0 ||
